+++ src/main/java/api/soldout/io/soldout/config/RoutingDataSourceConfig.java
@@ -7,12 +7,26 @@ import javax.sql.DataSource;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.jdbc.datasource.LazyConnectionDataSourceProxy;
+import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
+import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 /**
  * .
  */
 
 public class RoutingDataSourceConfig {
+
+  static class ReplicationRoutingDataSource extends AbstractRoutingDataSource {
+
+    @Override
+    protected Object determineCurrentLookupKey() {
+
+      return TransactionSynchronizationManager
+          .isCurrentTransactionReadOnly() ? DataSourceType.SLAVE : DataSourceType.MASTER;
+
+    }
+
+  }
 
   /**
    * .
