+++ src/main/java/api/soldout/io/soldout/config/SlaveMyBatisConfig.java
@@ -1,6 +1,5 @@
 package api.soldout.io.soldout.config;
 
-import com.zaxxer.hikari.HikariDataSource;
 import javax.sql.DataSource;
 import org.apache.ibatis.session.SqlSessionFactory;
 import org.mybatis.spring.SqlSessionFactoryBean;
@@ -8,12 +7,9 @@ import org.mybatis.spring.SqlSessionTemplate;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.jdbc.DataSourceBuilder;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
 
 /**
  * .
@@ -21,8 +17,8 @@ import org.springframework.context.annotation.Primary;
 
 @Configuration
 @MapperScan(value = " api.soldout.io.soldout.mapper",
-            sqlSessionFactoryRef = "masterSqlSessionFactory")
-public class MasterDataSourceConfig {
+            sqlSessionFactoryRef = "slaveSqlSessionFactory")
+public class SlaveMyBatisConfig {
 
   @Value("${mybatis.mapper-locations}")
   String mapperPath;
@@ -31,32 +27,18 @@ public class MasterDataSourceConfig {
    * .
    */
 
-  @Primary
-  @Bean(name = "masterDataSource")
-  @ConfigurationProperties(prefix = "spring.datasource.master.hikari")
-  public DataSource masterDataSource() {
-
-    return DataSourceBuilder.create()
-        .type(HikariDataSource.class)
-        .build();
-  }
-
-  /**
-   * .
-   */
-
-  @Primary
-  @Bean(name = "masterSqlSessionFactory")
-  public SqlSessionFactory sqlSessionFactory(@Qualifier("masterDataSource") DataSource dataSource,
-      ApplicationContext applicationContext) throws Exception {
+  @Bean(name = "slaveSqlSessionFactory")
+  public SqlSessionFactory sqlSessionFactory(@Qualifier("slaveDataSource") DataSource dataSource,
+      ApplicationContext applicationContext)
+      throws Exception {
     SqlSessionFactoryBean sqlSessionFactoryBean = new SqlSessionFactoryBean();
     sqlSessionFactoryBean.setDataSource(dataSource);
     sqlSessionFactoryBean.setMapperLocations(applicationContext.getResources(mapperPath));
     return sqlSessionFactoryBean.getObject();
   }
 
-  @Bean(name = "masterSessionTemplate")
-  public SqlSessionTemplate sqlSessionTemplate(@Qualifier("masterSqlSessionFactory")
+  @Bean(name = "slaveSessionTemplate")
+  public SqlSessionTemplate sqlSessionTemplate(@Qualifier("slaveSqlSessionFactory")
       SqlSessionFactory firstSqlSessionFactory) {
     return new SqlSessionTemplate(firstSqlSessionFactory);
   }
